release-manager-v2/config-crew-ai/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "config-crew-ai"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn",
    "crewai",
    "pydantic",
    "python-dotenv",
    "pyyaml"
]
